Preparing version 3.0.0
- refactor project structure.
- split decoders in its own files.
- Rename some decoders while keeping the original alias.
- Remove decoders: allOf, combine, isUndefined and isNull.
- Added decoders: null and undefined.
