+++ src/index.ts
@@ -1,2 +1,3 @@
-export { JsonDecoder, type FromDecoder } from './json-decoder';
-export { Ok, Err, type Result, err, ok } from './result';
+export * from './core';
+export * from './utils/result';
+export * from './schemas';
